Experimental (#1)
* Commit some tests

Those are somehow old tests on websocket.io. But we might use them at a later stage.

* Disconnect the worker in a cleaner way.

We create a connection class, which allows us to kill the worker.

* Script for testing serial ports.

* Update the readme

* Continue to clean up the connections

* Clean up the forms

We now have independent routes for all the forms. This way it is easier to understand

* Prepare to combine the connections into a single class

* moved the serial connection into a class

* made the transition to the new protocol

Started to move all the connection management into a single class.

* Got rid of the thread and thread_lock

* Improve connection

Moved all the background tasks into the protocol class

* Finishing up the connections

Made the behavior on connection more consistent. It feels rather straight-forward right now.