Dev (#6)
* Ability to remove the thermocontrol

* Added the cameras database

* Start to pull cameras from the database now.

* Ability to add and remove cameras from database

The background jobs do not work for the moment.

* Test if thread is still alive ...

* And I can trigger an image now.

* Removde the useless guppy class now.

* Add the tempcontrol table

* Move the thermocontrol to a database type now.

* Update the Arduino thermocontrol too.

* Also allow to view the details.

* Remove the tempcontrols entirely

Moved the tempcontrol entirely to the database model.

* Clean up the settings code a bit more.

* Automatically load the right parameters on the arduino on start up.

* Removed some useless rq dependency

* Fix an error on start-up

* Add the serial monitor database

* Make the serialmonitor databased

* Update the waittime of the serial monitor

* Add the email support

However, we have to save the login information on the PC. This has to be done properly, without comprimising passwords etc.

* Added a logging file.

* Kill off a funny text that does not mean anything.

* clean up the thermocontrol to raise clean errors.

* Get the values.

* plot string for the serialmonitor.

* save temp value

Added the ability to save the values of the tempcontrol. This could be improved at a later stage, but it is enough for testing with the NaLiControl

* fixed the tempcontrol ard_str

* Create simIndependentSerialPort.py

* Improved the README

* Pull the data from the serial

This is basically for the magnetometer control. This way we do not need the matlab script anymore.

* update the readme

* Save serialmonitor

Added the possibility to save the serialmonitor values to the hdf5

* Add the possibility of a modified baudrate

The different serial monitors can operate at different baudrates. Let us allow for changing them ...

* Update the baud_rate

We can now update the baud_rate if we need it.

* Update README.md