Dev (#18)
* Ability to remove the thermocontrol

* Added the cameras database

* Start to pull cameras from the database now.

* Ability to add and remove cameras from database

The background jobs do not work for the moment.

* Test if thread is still alive ...

* And I can trigger an image now.

* Removde the useless guppy class now.

* Add the tempcontrol table

* Move the thermocontrol to a database type now.

* Update the Arduino thermocontrol too.

* Also allow to view the details.

* Remove the tempcontrols entirely

Moved the tempcontrol entirely to the database model.

* Clean up the settings code a bit more.

* Automatically load the right parameters on the arduino on start up.

* Removed some useless rq dependency

* Fix an error on start-up

* Add the serial monitor database

* Make the serialmonitor databased

* Update the waittime of the serial monitor

* Add the email support

However, we have to save the login information on the PC. This has to be done properly, without comprimising passwords etc.

* Added a logging file.

* Kill off a funny text that does not mean anything.

* clean up the thermocontrol to raise clean errors.

* Get the values.

* plot string for the serialmonitor.

* save temp value

Added the ability to save the values of the tempcontrol. This could be improved at a later stage, but it is enough for testing with the NaLiControl

* fixed the tempcontrol ard_str

* Create simIndependentSerialPort.py

* Improved the README

* Pull the data from the serial

This is basically for the magnetometer control. This way we do not need the matlab script anymore.

* update the readme

* Save serialmonitor

Added the possibility to save the serialmonitor values to the hdf5

* Add the possibility of a modified baudrate

The different serial monitors can operate at different baudrates. Let us allow for changing them ...

* Update the baud_rate

We can now update the baud_rate if we need it.

* Update README.md

* Start the web arduino

Start the web control class

* Add possibility of a webarduino

* Start communicating with arduino

* Simple arduino readout

* Start and stop the web arduino.

No updated settings yet.

* Update change_arduino.html

Get start and stop running

* Update models.py

Start to use an abstract underlying class for the databases...

* Update models.py

* Start extending templates

Start to use more templates for the different sites.

* Start to read out the response from the Arduino

* Only listen for a webarduino

* finish the monitoring

It should work separately now. Only the settings are let to be updated.

* Communicate properly with the arduino

* Update models.py

* Get first data from the arduino

* Update models.py

* Update requirements.txt

* Delete simIndependentSerialPort.py

* Get the right form.

* Update handlers.py

* Fixed the readout for the analog sensors again...

* Flask factory (#9)

* Started to remove the global app variable

So I think that I removed it, but this now means that I have get the app context right. For that we need to call the right settings module.

* Update tests.py

Set up the tests in a cleaner env.

* Make the do_work run rather properly now.

* Update models.py

* Heroku test (#10)

* Prepare for heroku

* Update Procfile

* Update Procfile

* Change name

* Get the server a bit better.

* git fix

Remove some git fancyness, that was just a toy anyway ...

* SimpleAccessToTheYun

* Start moving the arduino program towards the Yun.

* Update README.md

* Add a model for pulling the temperature

So we can pull the temperature right now, but we really want to pull a whole website.

* Get the Arduino to answer to temp requests

* New Yun API

* Docker tests (#11)

* Prepare for docker

* Make a dockerfile that compiles.

So it compiles, but I am not able to start it and I had to remove a lot of dependencies ...

* Clean up for docker

Now I can run it with a simple

docker run -p 8000:5000 devicecontrolserver:latest

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Clean up the removal process

* Make the docker version also compatible with heroku

* Update README.md

* Update models.py

* Update requirements.txt

* Make the thread_id a string

Maybe this avoids the heroku overflow problems

* Update .travis.yml

* Updated the YUn Arduino program

* Fix the serial port

Basically we had an undefined variable for the serials that we removed as it was unused.

* Update YunPutAndGet.ino

* Update the Arduino code

* Add users to make it possible to have multiple machines (#13)

* Setting up the login form, which is not running for the moment

* Have now a login page.

Not able to register yet, but we can login.

* Login and Logout is running now.

* Add the device owner

Make it possible to set owners for a device.

* Make control only available for logged in users

* New devices now have an owner.

* Hide the arduinos more and more

* Deny access for non-identified users.

* Update __init__.py

* Bootstrap4 (#15)

* Nice header

* Clean the index file and make the log file cute.

* removed most wtf dependencies that I could find for the moment.

* Reactive (#16)

* The open and close is now reactive

* Close everywhere

Update everywhere that the connection is closed.

* Show the open status too.

* Also updates when it is open now...

* Add labels for improved reactivity

* I actually broke the system with the new bootstrap

Updating to the different bootstrap broke all my forms. So we need tests (!!!!)

* Fix forms (#17)

* Undo the transition to bootstrap 4

* Fixed the index a bit.

* Add a first test to avoid breaking everything all the time.

* Update .travis.yml

* Fix some merge problems